Refactored out AddModal, EditModal and DeleteModal
Into separate files, included them in DiskPanel and fixed tests as well.
